requirements.txt
Flask==1.1.2
numpy
torch
torchvision
dominate>=2.4.0
visdom>=0.1.8.8
scipy
Pillow
